requirements.txt
tensorflow-gpu
keras
deepdish
scipy
numpy
pandas
matplotlib
ipykernel
